refactor: Rename spinners to spinner
